+++ src/plugins/dev-mode/error-messages.ts
@@ -148,6 +148,8 @@ export const ERROR_MESSAGES: { [k: string]: string } = {
     SC24: 'SchemaCheck: required fields must be set via array. See https://spacetelescope.github.io/understanding-json-schema/reference/object.html#required',
     SC25: 'SchemaCheck: compoundIndexes needs to be specified in the indexes field',
     SC26: 'SchemaCheck: indexes needs to be specified at collection schema level',
+    SC27: 'SchemaCheck: encrypted fields need to be specified at collection schema level',
+    SC28: 'SchemaCheck: encrypted fields is not defined in the schema',
 
     // plugins/validate.js
     VD1: 'Sub-schema not found, does the schemaPath exists in your schema?',